[fix] グッズ作成時にタグと購入場所を配列に変換するように修正, コレクションリスト内の画像ビューのリンクの場違いを修正
