+++ frontend/src/app/api/itemDetail/route.ts
@@ -1,6 +1,4 @@
-import {
-  BACKEND_UPDATE_ITEM_KEYS as keys
-} from "@/constants"
+import formToJson from "@/utils/formToJson"
 import makeToken from "@/utils/makeToken"
 import { NextRequest } from "next/server"
 
@@ -53,21 +51,7 @@ export async function POST(
   if (!itemId)  return Response.error()
 
   const requestUrl = `${ backendUrl }items/${ itemId }`
-  const arrayTypeList = [keys.tags, keys.retailers]
-  const formToObjectToArray = Object.entries(Object.fromEntries(formData.entries()))
-  const newObject: {[key: string]: string | string[]} = {}
-  formToObjectToArray.forEach(ary => {
-    const key = ary[0]
-    const value = ary[1].toString()
-    if (value !== "") {
-      if (arrayTypeList.includes(key)) {
-        newObject[key] = value.split(",")
-      } else {
-        newObject[key] = value
-      }
-    }
-  })
-  const jsonData = JSON.stringify(newObject)
+  const jsonData = formToJson(formData)
 
   const response = await fetch(
     requestUrl,
