+++ frontend/src/app/api/items/create/route.ts
@@ -1,3 +1,4 @@
+import formToJson from "@/utils/formToJson"
 import makeToken from "@/utils/makeToken"
 import { NextRequest } from "next/server"
 
@@ -49,7 +50,7 @@ export async function POST(
 
   const formData = await request.formData()
   const requestUrl = `${ backendUrl }items`
-  const json = JSON.stringify(Object.fromEntries(formData.entries()))
+  const json = formToJson(formData)
 
   const response = await fetch(
     requestUrl,
